Added features to get a multithread RT app.
I see a loong and lonesome road ahead of me. Also. Who  in the Hell is the
one that writes the copyright rules for Youtube? Seriously.

However, as you might see, this application uses a great deal of
dereferenced structures and classes. It's the first time that I have used
C++, so I don't really know if I am doing this right... :3