Threads working! Mutexes seem to work... :3
The only problem is this, now. As you might have seen ALSA has an old
function to get the data via mmap. I've taken it from the microphone.c
sample program I have. It works this way. I get the data and I manage to
output it directly on the terminal. The problem arises when I try to modify
it to put this data inside the shared zone.

No, it doesn't break because I try to write in the shared memory zone. It
breaks even before! I get a Segmentation Fault even BEFORE using the
pointers. I cannot understand why I get these results... Every time I touch
that function it breaks in points where are used variables I don't modify!
It's unbelievable. Well, I'll manage to fix it tomorrow.