Guess What? (￣▽￣; )
At last! We have something that works. Or so it seems... Now it really is a
Real Time Application. I mean, it also crashes "Oh! So nicely" when someone
tries to make it run without amministrative permissions XD
Oh, well. Time to do some butchery to remove all the test functions, the
weird comments (That you will be able to see here on GitHub, I suppose ;)
), the strange things everywhere inside each file and then write some good
documentation.